refactor: make completer cli lazy loaded
